refactor: move all tests to the root of test folder
